Day 12 revised solution
Reimplementation of the original recursive solution which ended up being
deleted and replaced with the current "original" Day 12 solution when a
problem unrelated to its recursive nature caused stack overflows. This
implementation is slightly different in its use of a custom `Smalls`
type to keep track of which small caves have already been visited, as
opposed to the original recursive solution, and current "original"
solution, which use `HashSet`s.